Big Update to main README.md
Lots of changes.

PhaseLink is now a protocol, not a device. It is called the PhaseLink Transmission Model.

QSSD is a data storage device with protocols. It relies on PhaseLink to function, and PhaseLink is integrated into the system, which makes implementation incredibly complex.

Quelephone is a device with protocols. It relies on PhaseLink to function, and PhaseLink technology is integrated into the system, which makes implementation incredibly complex, though the system itself is for communication.